Refactor event fetching in SentryClient to use project slug instead of ID
- Update events method to retrieve project slug from the projects list
- Add logging for cases where the project slug cannot be found
- Enhance query construction for event fetching with the project slug

We identified that the events method in the SentryClient class was using the project ID directly in the URL instead of the project slug.
We modified the events method to look up the project slug based on the project ID and use that in the URL.
We made sure to handle the case where the project slug might not be found.
These changes should allow the tap to correctly fetch events from the Sentry API. The next time you run the tap, it should be able to sync events data.